+++ bluepysnap/simulation.py
@@ -16,12 +16,14 @@
 # 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 """Simulation access."""
 
+import warnings
 from pathlib import Path
 
 from cached_property import cached_property
 
 from bluepysnap.config import SimulationConfig
 from bluepysnap.exceptions import BluepySnapError
+from bluepysnap.input import get_simulation_inputs
 from bluepysnap.node_sets import NodeSets
 
 
@@ -44,6 +46,16 @@ def _collect_frame_reports(sim):
 
         res[name] = cls(sim, name)
     return res
+
+
+def _warn_on_overwritten_node_sets(overwritten, print_max=10):
+    """Helper function to warn and print overwritten nodesets."""
+    if (n := len(overwritten)) > 0:
+        names = ", ".join(list(overwritten)[:print_max]) + (", ..." if n > print_max else "")
+        warnings.warn(
+            f"Simulation node sets overwrite {n} node set(s) in Circuit node sets: {names}",
+            RuntimeWarning,
+        )
 
 
 class Simulation:
@@ -86,6 +98,11 @@ class Simulation:
         return self.to_libsonata.output
 
     @property
+    def inputs(self):
+        """Access the inputs section."""
+        return get_simulation_inputs(self.to_libsonata)
+
+    @property
     def run(self):
         """Access to the complete run dictionary for this simulation."""
         return self.to_libsonata.run
@@ -125,8 +142,16 @@ class Simulation:
     @cached_property
     def node_sets(self):
         """Returns the NodeSets object bound to the simulation."""
-        path = self.to_libsonata.node_sets_file
-        return NodeSets.from_file(path) if path else NodeSets.from_dict({})
+        try:
+            node_sets = NodeSets.from_file(self.circuit.to_libsonata.node_sets_path)
+        except BluepySnapError:  # Error raised if circuit can not be instantiated
+            node_sets = NodeSets.from_dict({})
+
+        if path := self.to_libsonata.node_sets_file:
+            overwritten = node_sets.update(NodeSets.from_file(path))
+            _warn_on_overwritten_node_sets(overwritten)
+
+        return node_sets
 
     @cached_property
     def spikes(self):